Refactor: 전적 가져오기 -> 5개 가져오기
전적 1개 가져오는 것에서 5개 가져오는 것으로 변경
