+++ Search-service/src/main/java/com/ggwp/searchservice/match/MatchRepository/MatchRepository.java
@@ -3,5 +3,7 @@ package com.ggwp.searchservice.match.MatchRepository;
 import com.ggwp.searchservice.match.domain.Match;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MatchRepository extends JpaRepository<Match,Long> {
+public interface MatchRepository extends JpaRepository<Match, Long> {
+
+    boolean existsByMatchId(String matchId);
 }
