+++ Search-service/src/main/java/com/ggwp/searchservice/match/MatchRepository/MatchSummonerRepository.java
@@ -1,0 +1,8 @@
+package com.ggwp.searchservice.match.MatchRepository;
+
+import com.ggwp.searchservice.match.domain.MatchSummoner;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface MatchSummonerRepository extends JpaRepository<MatchSummoner, Long> {
+    
+}
