+++ Search-service/src/main/java/com/ggwp/searchservice/match/feign/LoLToMatchFeign.java
@@ -7,6 +7,8 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestParam;
 
+import java.util.List;
+
 @FeignClient(name = "lol-match", url = "https://asia.api.riotgames.com/lol/")
 public interface LoLToMatchFeign {
 
@@ -15,13 +17,12 @@ public interface LoLToMatchFeign {
             @PathVariable("matchId") String matchId,
             @RequestParam("api_key") String apiKey
     );
-//    @GetMapping("match/v5/matches/by-puuid/{puuid}/ids")
-//    List<String> getMatchId(
-//            @PathVariable("puuid") String puuid,
-//            @RequestParam("api_key") String apiKey
-//    );
 
+    @GetMapping("match/v5/matches/by-puuid/{puuid}/ids?start=0&count=5")
+    List<String> getMatchIds(
+            @PathVariable("puuid") String puuid,
+            @RequestParam("api_key") String apiKey
+    );
 
-    // https://asia.api.riotgames.com/lol/match/v5/matches/KR_6679031558?api_key=RGAPI-4cac8adf-0eac-4db1-80ef-d799850cb1a7
 }
