+++ Search-service/src/main/java/com/ggwp/searchservice/summoner/domain/Summoner.java
@@ -3,15 +3,13 @@ package com.ggwp.searchservice.summoner.domain;
 import com.ggwp.searchservice.league.domain.League;
 import com.ggwp.searchservice.match.domain.Match;
 import com.ggwp.searchservice.summoner.dto.ResponseGetSummonerDto;
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.Id;
-import jakarta.persistence.OneToMany;
+import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+import java.util.ArrayList;
 import java.util.List;
 
 @Entity
@@ -35,18 +33,17 @@ public class Summoner {
 
     private Long revisionDate; // 전적 업데이트 날짜
 
-    private Long summonerLevel; // 소환사 레벨
+    private int summonerLevel; // 소환사 레벨
 
     // 리그 연결 일대다 ( 소환사 1 : 리그 2)
     @OneToMany(mappedBy = "summoner")
     private List<League> leagues;
 
-//    // 매치 연결 일대다 ( 소환사 1 : 매치 10)
-//    @OneToMany(mappedBy = "summoner")
-//    private List<Match> matches;
+    @ManyToMany(mappedBy = "summoners")
+    private List<Match> matches = new ArrayList<>();
 
     // 소환사 Entity를 -> DTO로 변경
-    public ResponseGetSummonerDto toDto(Summoner summoner){
+    public ResponseGetSummonerDto toDto(Summoner summoner) {
         return ResponseGetSummonerDto.builder()
                 .id(summoner.getId())
                 .profileIconId(summoner.getProfileIconId())
@@ -56,5 +53,13 @@ public class Summoner {
                 .revisionDate(summoner.getRevisionDate())
                 .summonerLevel(summoner.getSummonerLevel())
                 .build();
+    }
+
+    public void addMatch(Match match) {
+        if (this.matches == null) {
+            this.matches = new ArrayList<>();
+        }
+        this.matches.size(); // Lazy 때문에 해야함
+        this.matches.add(match);
     }
 }
