+++ Search-service/src/main/resources/application.yml
@@ -1,37 +1,6 @@
-# spring
-  spring:
-    application:
-      name: Search-service
-    profiles:
-      active: db
-    datasource:
-      driver-class-name: ${driver-class-name}
-      url: ${url}
-      username: ${username_}
-      password: ${password}
-  #    kafka:
-  #      bootstrap-servers: localhost:9092
-
-    # jpa
-    jpa:
-      hibernate:
-        ddl-auto: create
-      properties:
-        hibernate:
-          format_sql: true
-          show_sql: true
-# rabbitmq
-#  rabbitmq:
-#    host: localhost
-#    username: guest
-#    password: guest
-#    port: 5672
-#    eureka
-#  eureka:
-#    instance:
-#      instance-id: ${spring.application.name}:${spring.application.instance_id:${random.value}}
-#    client:
-#      register-with-eureka: true
-#      fetch-registry: true
-#      service-url:
-#        defaultZone: http://127.0.0.1:8761/eureka
+spring:
+  profiles:
+    active:
+      - dev
+    include:
+      - db
